Fix EEPROM SD

+++ MK4duo/src/sd/cardreader.h
@@ -191,8 +191,8 @@
       #if HAS_EEPROM_SD
         static void open_eeprom_sd(const bool read);
         static void close_eeprom_sd();
-        FORCE_INLINE static int16_t write_eeprom_data(void* buf, uint16_t nbyte)  { return eeprom_file.isOpen() ? eeprom_file.write(buf, nbyte) : -1; }
-        FORCE_INLINE static int16_t read_eeprom_data(void* buf, uint16_t nbyte)   { return eeprom_file.isOpen() ? eeprom_file.read(buf, nbyte) : -1; }
+        FORCE_INLINE static int write_eeprom_data(void* buf, uint16_t nbyte)  { return eeprom_file.isOpen() ? eeprom_file.write(buf, nbyte) : -1; }
+        FORCE_INLINE static int read_eeprom_data(void* buf, uint16_t nbyte)   { return eeprom_file.isOpen() ? eeprom_file.read(buf, nbyte) : -1; }
       #endif
 
       #if ENABLED(SD_SETTINGS)
